+++ great_expectations/render/renderer/content_block/__init__.py
@@ -1,7 +1,8 @@
 from .value_list_content_block import ValueListContentBlockRenderer
 from .table_content_block import TableContentBlockRenderer
 from .bullet_list_content_block import (
-    PrescriptiveBulletListContentBlockRenderer,
-    PrescriptiveEvrTableContentBlockRenderer
+    ExpectationSuiteBulletListContentBlockRenderer
 )
+from great_expectations.render.renderer.content_block.validation_results_table_content_block import \
+    ValidationResultsTableContentBlockRenderer
 from .exception_list_content_block import ExceptionListContentBlockRenderer
