+++ great_expectations/render/renderer/page_renderer.py
@@ -5,30 +5,22 @@ import pypandoc
 
 from .renderer import Renderer
 from .column_section_renderer import (
-    DescriptiveColumnSectionRenderer,
-    PrescriptiveColumnSectionRenderer,
+    ProfilingResultsColumnSectionRenderer,
+    ExpectationSuiteColumnSectionRenderer,
+    ValidationResultsColumnSectionRenderer
 )
 from .other_section_renderer import (
-    DescriptiveOverviewSectionRenderer,
+    ProfilingResultsOverviewSectionRenderer,
 )
 
 
-class PrescriptivePageRenderer(Renderer):
-
+class ValidationResultsPageRenderer(Renderer):
     @classmethod
-    def render(cls, expectations={}, validation_results={}):
-        if expectations:
-            return cls._render_expectation_suite(expectations)
-        else:
-            return cls._render_validation(validation_results)
-            
-    @classmethod
-    def _render_validation(cls, validation_results):
+    def render(cls, validation_results={}):
         run_id = validation_results['meta']['run_id']
         full_data_asset_identifier = validation_results['meta']['data_asset_name'] or ""
         expectation_suite_name = validation_results['meta']['expectation_suite_name']
         short_data_asset_name = full_data_asset_identifier.split('/')[-1]
-        ge_version = validation_results["meta"]["great_expectations.__version__"]
     
         # Group EVRs by column
         columns = {}
@@ -44,69 +36,108 @@ class PrescriptivePageRenderer(Renderer):
     
         ordered_columns = Renderer._get_column_list_from_evrs(validation_results)
     
+        overview_content_blocks = [
+            cls._render_validation_header(),
+            cls._render_validation_info(validation_results=validation_results),
+            cls._render_validation_statistics(validation_results=validation_results)
+        ]
+    
         if "data_asset_name" in validation_results["meta"] and validation_results["meta"]["data_asset_name"]:
             data_asset_name = short_data_asset_name
         else:
             data_asset_name = None
-
+    
         sections = [
             {
                 "section_name": "Overview",
-                "content_blocks": [
-                    {
-                        "content_block_type": "header",
-                        "header": "Validation Overview",
-                        "styling": {
-                            "classes": ["col-12"],
-                            "header": {
-                                "classes": ["alert", "alert-secondary"]
-                            }
-                        }
-                    },
-                    {
-                        "content_block_type": "table",
-                        "header": "Info",
-                        "table": [
-                            ["Full Data Asset Identifier", full_data_asset_identifier],
-                            ["Expectation Suite Name", expectation_suite_name],
-                            ["Great Expectations Version", ge_version],
-                            ["Run ID", run_id]
-                        ],
-                        "styling": {
-                            "classes": ["col-12", "table-responsive"],
-                            "styles": {
-                                "margin-top": "20px"
-                            },
-                            "body": {
-                                "classes": ["table", "table-sm"]
-                            }
-                        },
-                    }
-                ]
+                "content_blocks": overview_content_blocks
             }
         ]
     
         sections += [
-                    PrescriptiveColumnSectionRenderer.render(
-                        validation_results=columns[column],
-                    ) for column in ordered_columns
-                ]
+            ValidationResultsColumnSectionRenderer.render(
+                validation_results=columns[column],
+            ) for column in ordered_columns
+        ]
     
         return {
-            "renderer_type": "PrescriptivePageRenderer",
+            "renderer_type": "ValidationResultsColumnSectionRenderer",
             "data_asset_name": data_asset_name,
             "full_data_asset_identifier": full_data_asset_identifier,
             "page_title": run_id + "-" + expectation_suite_name + "-Descriptive",
             "sections": sections
         }
-        
+    
     @classmethod
-    def _render_expectation_suite(cls, expectations):
+    def _render_validation_header(cls):
+        return {
+            "content_block_type": "header",
+            "header": "Validation Overview",
+            "styling": {
+                "classes": ["col-12"],
+                "header": {
+                    "classes": ["alert", "alert-secondary"]
+                }
+            }
+        }
+    
+    @classmethod
+    def _render_validation_info(cls, validation_results):
+        run_id = validation_results['meta']['run_id']
+        full_data_asset_identifier = validation_results['meta']['data_asset_name'] or ""
+        expectation_suite_name = validation_results['meta']['expectation_suite_name']
+        ge_version = validation_results["meta"]["great_expectations.__version__"]
+        success = validation_results["success"]
+        
+        return {
+            "content_block_type": "table",
+            "header": "Info",
+            "table": [
+                ["Full Data Asset Identifier", full_data_asset_identifier],
+                ["Expectation Suite Name", expectation_suite_name],
+                ["Great Expectations Version", ge_version],
+                ["Run ID", run_id],
+                ["Validation Succeeded", success]
+            ],
+            "styling": {
+                "classes": ["col-12", "table-responsive"],
+                "styles": {
+                    "margin-top": "20px"
+                },
+                "body": {
+                    "classes": ["table", "table-sm"]
+                }
+            },
+        }
+    
+    @classmethod
+    def _render_validation_statistics(cls, validation_results):
+        statistics = validation_results["statistics"]
+        return {
+            "content_block_type": "table",
+            "header": "Statistics",
+            "table": [
+                ["Evaluated Expectations", statistics["evaluated_expectations"]],
+                ["Successful Expectations", statistics["successful_expectations"]],
+                ["Unsuccessful Expectations", statistics["unsuccessful_expectations"]],
+                ["Success Percent", "{0:.2f}%".format(statistics["success_percent"])],
+            ],
+            "styling": {
+                "classes": ["col-6", "table-responsive"],
+                "styles": {
+                    "margin-top": "20px"
+                },
+                "body": {
+                    "classes": ["table", "table-sm"]
+                }
+            },
+        }
+
+
+class ExpectationSuitePageRenderer(Renderer):
+    @classmethod
+    def render(cls, expectations={}):
         full_data_asset_identifier = expectations.get("data_asset_name") or ""
-        short_data_asset_name = full_data_asset_identifier.split('/')[-1]
-        data_asset_type = expectations.get("data_asset_type")
-        expectation_suite_name = expectations.get("expectation_suite_name")
-        ge_version = expectations["meta"]["great_expectations.__version__"]
         
         # Group expectations by column
         columns = {}
@@ -123,17 +154,17 @@ class PrescriptivePageRenderer(Renderer):
             if column not in columns:
                 columns[column] = []
             columns[column].append(expectation)
-
+            
             # if possible, get the order of columns from expect_table_columns_to_match_ordered_list
             if expectation["expectation_type"] == "expect_table_columns_to_match_ordered_list":
                 exp_column_list = expectation["kwargs"]["column_list"]
                 if exp_column_list and len(exp_column_list) > 0:
                     ordered_columns = exp_column_list
-
+        
         # if no order of columns is expected, sort alphabetically
         if not ordered_columns:
             ordered_columns = sorted(list(columns.keys()))
-
+        
         overview_content_blocks = [
             cls._render_asset_header(expectations),
             cls._render_asset_info(expectations)
@@ -144,27 +175,27 @@ class PrescriptivePageRenderer(Renderer):
             overview_content_blocks.append(asset_notes_content_block)
             # import json
             # print(json.dumps(overview_content_blocks, indent=2))
-
+        
         sections = [
             {
                 "section_name": "Overview",
                 "content_blocks": overview_content_blocks,
             }
         ]
-
+        
         sections += [
-                PrescriptiveColumnSectionRenderer.render(expectations=columns[column]) for column in ordered_columns
-            ]
-
+            ExpectationSuiteColumnSectionRenderer.render(expectations=columns[column]) for column in ordered_columns
+        ]
+        
         return {
-            # "renderer_type": "PrescriptivePageRenderer",
+            # "renderer_type": "ExpectationSuitePageRenderer",
             # "data_asset_name": short_data_asset_name,
             "full_data_asset_identifier": full_data_asset_identifier,
             "page_title": expectation_suite_name,
             "utm_medium": "prescriptive-expectation-suite-page",
             "sections": sections
         }
-
+    
     @classmethod
     def _render_asset_header(cls, expectations):
         return {
@@ -177,16 +208,14 @@ class PrescriptivePageRenderer(Renderer):
                 }
             }
         }
-
+    
     @classmethod
     def _render_asset_info(cls, expectations):
-
         full_data_asset_identifier = expectations.get("data_asset_name") or ""
-        short_data_asset_name = full_data_asset_identifier.split('/')[-1]
         data_asset_type = expectations.get("data_asset_type")
         expectation_suite_name = expectations.get("expectation_suite_name")
         ge_version = expectations["meta"]["great_expectations.__version__"]
-
+        
         return {
             "content_block_type": "table",
             "header": "Info",
@@ -206,15 +235,15 @@ class PrescriptivePageRenderer(Renderer):
                 }
             },
         }
-
+    
     @classmethod
     def _render_asset_notes(cls, expectations):
-
+        
         content = []
-
+        
         if "expectations" in expectations:
-            #This if statement is a precaution in case the expectation suite doesn't contain expectations.
-            #Once we have more strongly typed classes for suites, this shouldn't be necessary.
+            # This if statement is a precaution in case the expectation suite doesn't contain expectations.
+            # Once we have more strongly typed classes for suites, this shouldn't be necessary.
             
             total_expectations = len(expectations["expectations"])
             columns = []
@@ -222,9 +251,9 @@ class PrescriptivePageRenderer(Renderer):
                 if "column" in exp["kwargs"]:
                     columns.append(exp["kwargs"]["column"])
             total_columns = len(set(columns))
-
+            
             content = content + [
-                #TODO: Leaving these two paragraphs as placeholders for later development.
+                # TODO: Leaving these two paragraphs as placeholders for later development.
                 # "This Expectation suite was first generated by {BasicDatasetProfiler} on {date}, using version {xxx} of Great Expectations.",
                 # "{name}, {name}, and {name} have also contributed additions and revisions.",
                 "This Expectation suite currently contains %d total Expectations across %d columns." % (
@@ -232,17 +261,17 @@ class PrescriptivePageRenderer(Renderer):
                     total_columns,
                 ),
             ]
-
+        
         if "notes" in expectations["meta"]:
             notes = expectations["meta"]["notes"]
             note_content = None
-
+            
             if type(notes) == str:
                 note_content = [notes]
-
+            
             elif type(notes) == list:
                 note_content = notes
-
+            
             elif type(notes) == dict:
                 if "format" in notes:
                     if notes["format"] == "string":
@@ -252,25 +281,26 @@ class PrescriptivePageRenderer(Renderer):
                             note_content = notes["content"]
                         else:
                             logger.warning("Unrecognized Expectation suite notes format. Skipping rendering.")
-                        
+                    
                     elif notes["format"] == "markdown":
-                        #???: Should converting to markdown be the renderer's job, or the view's job?
-                        #Renderer is easier, but will end up mixing HTML strings with content_block info.
+                        # ???: Should converting to markdown be the renderer's job, or the view's job?
+                        # Renderer is easier, but will end up mixing HTML strings with content_block info.
                         if type(notes["content"]) == str:
                             note_content = [pypandoc.convert_text(notes["content"], format='md', to="html")]
-
+                        
                         elif type(notes["content"]) == list:
-                            note_content = [pypandoc.convert_text(note, format='md', to="html") for note in notes["content"]]
-
+                            note_content = [pypandoc.convert_text(note, format='md', to="html") for note in
+                                            notes["content"]]
+                        
                         else:
                             logger.warning("Unrecognized Expectation suite notes format. Skipping rendering.")
-
+                
                 else:
                     logger.warning("Unrecognized Expectation suite notes format. Skipping rendering.")
-
+            
             if note_content != None:
                 content = content + note_content
-
+        
         return {
             "content_block_type": "text",
             "header": "Notes",
@@ -287,7 +317,7 @@ class PrescriptivePageRenderer(Renderer):
         }
 
 
-class DescriptivePageRenderer(Renderer):
+class ProfilingResultsPageRenderer(Renderer):
 
     @classmethod
     def render(cls, validation_results):
@@ -309,7 +339,7 @@ class DescriptivePageRenderer(Renderer):
             columns[column].append(evr)
 
         ordered_columns = Renderer._get_column_list_from_evrs(validation_results)
-        column_types = DescriptiveOverviewSectionRenderer._get_column_types(validation_results)
+        column_types = ProfilingResultsOverviewSectionRenderer._get_column_types(validation_results)
 
         if "data_asset_name" in validation_results["meta"] and validation_results["meta"]["data_asset_name"]:
             data_asset_name = short_data_asset_name
@@ -317,20 +347,20 @@ class DescriptivePageRenderer(Renderer):
             data_asset_name = None
 
         return {
-            "renderer_type": "DescriptivePageRenderer",
+            "renderer_type": "ProfilingResultsPageRenderer",
             "data_asset_name": data_asset_name,
             "full_data_asset_identifier": full_data_asset_identifier,
             "page_title": run_id + "-" + expectation_suite_name + "-Descriptive",
             "utm_medium": "descriptive-validation-page",
             "sections":
                 [
-                    DescriptiveOverviewSectionRenderer.render(
+                    ProfilingResultsOverviewSectionRenderer.render(
                         validation_results,
                         section_name="Overview"
                     )
                 ] +
                 [
-                    DescriptiveColumnSectionRenderer.render(
+                    ProfilingResultsColumnSectionRenderer.render(
                         columns[column],
                         section_name=column,
                         column_type=column_types.get(column),
